ステータス関連を別ファイルに分離
